modularize the configuration
