+++ ccproxy/CLAUDE.md
@@ -10,3 +10,22 @@ Guidelines:
 - Do not log secrets; use ccproxy.logging helpers
 - Preserve UTF-8 when handling bytes/strings
 - JSON logging automatically omits null values for cleaner output
+
+## New Modules Added
+
+### ccproxy/constants.py
+
+Contains all constant values used throughout the application:
+
+- UTF-8 enforcement message for models that support developer role
+- Model capability sets (reasoning effort, temperature, developer message support)
+- Top-tier models for OpenAI and Anthropic
+- Input and output token limits for all supported models
+
+### ccproxy/enums.py
+
+Contains all enumeration classes used throughout the application:
+
+- MessageRoles enum (Developer, System, User)
+- ReasoningEfforts enum (High, Medium, Low)
+- TruncationStrategy and TruncationConfig for message truncation handling
